Fix off-by-one bug when calculating the bundle size
Remember position in parcel before adding bundle signature. The old code supposed to
work by chance: the resolved type was NULL (so only -1 as int32 was stored in
the parcel) and redundant -1 (as int32) was added after the resolved type.
Thus, the former was actually a null string in the bundle and the latter acted the part
of the (null) resolved type.

Don't put garbage at the end of intent (after sticky which is supposed to be
the last parameter in an intent). Again, it seemed to work before due to Android stops
to parse intent just after it gets all required parameters.